+++ packages/components/src/components/flyout/flyout.tsx
@@ -9,6 +9,7 @@ import {
 import { getComponentCss } from './flyout-styles';
 import {
   AllowedTypes,
+  applyConstructableStylesheetStyles,
   attachComponentCss,
   getHasConstructableStylesheetSupport,
   getIntersectionObserverStickyArea,
@@ -25,6 +26,7 @@ import {
   warnIfDeprecatedPropValueIsUsed,
 } from '../../utils';
 import type { PropTypes, SelectedAriaAttributes, Theme } from '../../types';
+import { getSlottedAnchorStyles } from '../../styles';
 
 const propTypes: PropTypes<typeof Flyout> = {
   open: AllowedTypes.boolean,
@@ -69,6 +71,10 @@ export class Flyout {
 
   public componentShouldUpdate(newVal: unknown, oldVal: unknown): boolean {
     return hasPropValueChanged(newVal, oldVal);
+  }
+
+  public connectedCallback(): void {
+    applyConstructableStylesheetStyles(this.host, getSlottedAnchorStyles);
   }
 
   public componentWillRender(): void {