+++ packages/components/CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ### [Unreleased]
 
+### [3.15.0] - 2024-05-16
+
+### [3.15.0-rc.5] - 2024-05-16
+
 #### Added
 
 - `Flyout`:
@@ -45,6 +49,28 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 - `Button Pure`: Prop `underline` to show an underline for the label
   ([#3212](https://github.com/porsche-design-system/porsche-design-system/pull/3212))
+- Partials: Added new option `globalStyles` to `getInitialStyles` to disable global reset styles.  
+  ([#3213](https://github.com/porsche-design-system/porsche-design-system/pull/3213))
+
+#### Changed
+
+- Banner: Refactor Banner to use native `popover`
+  ([#3196](https://github.com/porsche-design-system/porsche-design-system/pull/3196))
+- Partials: `getInitialStyles` only contain hydration visibility and global styles. All other styles are handled by
+  constructable stylesheets at component level.
+  ([#3213](https://github.com/porsche-design-system/porsche-design-system/pull/3213))
+- `Table`: Removed slotted image style `verticalAlign: 'middle'` from initialStyles
+  ([#3213](https://github.com/porsche-design-system/porsche-design-system/pull/3213))
+- `Tabs Bar`: Removed sibling tabpanel focus style from initialStyles
+  ([#3213](https://github.com/porsche-design-system/porsche-design-system/pull/3213))
+
+#### Fixed
+
+- `Carousel`: Accessible name of carousel region wrapper
+  ([#3220](https://github.com/porsche-design-system/porsche-design-system/pull/3220))
+- `aria` property now supports escaped single quotes inside JSON strings, e.g.
+  `aria="{ 'aria-label': 'You can\'t do that? yes you can!' }"`
+  ([#3217](https://github.com/porsche-design-system/porsche-design-system/pull/3217))
 
 ### [3.15.0-rc.4] - 2024-05-06
 
@@ -109,7 +135,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 - `Model Signature`: Enabling the use of hex colors, CSS gradients, CSS image and video masks. In addition, the size was
   slightly adjusted. ([#3153](https://github.com/porsche-design-system/porsche-design-system/pull/3153))
 
-### Fixed
+#### Fixed
 
 - `Modal`: Missing box-shadow on sticky footer when slotted content changes
   ([#3154](https://github.com/porsche-design-system/porsche-design-system/pull/3154))